requirements.txt
aif360==0.6.1
aiohttp==3.8.5
aiosignal==1.3.1
aiosqlite==0.19.0
antlr4-python3-runtime==4.9.3
anyio==3.7.1
appdirs==1.4.4
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
arrow==1.2.3
asttokens==2.2.1
async-lru==2.0.4
async-timeout==4.0.3
attrs==23.1.0
auto-sklearn==0.15.0
Babel==2.12.1
backcall==0.2.0
beautifulsoup4==4.12.2
bleach==6.0.0
blitz-bayesian-pytorch==0.2.7
botorch==0.8.1
catboost==1.2.2
category-encoders==2.6.4
causal-learn==0.1.3.8
certifi==2023.7.22
cffi==1.15.1
cfgv==3.4.0
charset-normalizer==3.2.0
clarabel==0.7.1
click==8.1.7
cloudpickle==2.2.0
cmake==3.27.2
colorful==0.5.6
colorlover==0.3.0
comm==0.1.4
ConfigSpace==0.4.21
contourpy==1.0.6
cvxpy==1.4.2
cycler==0.11.0
cymem==2.0.10
Cython==3.0.0
dask==2023.10.0
dataclasses-json==0.5.14
debtcollector==2.5.0
debugpy==1.6.7.post1
decorator==5.1.1
defusedxml==0.7.1
distlib==0.3.7
distributed==2023.10.0
docker-pycreds==0.4.0
dowhy==0.11.1
ecos==2.0.13
einops==0.6.1
emcee==3.1.4
exceptiongroup==1.0.1
executing==1.2.0
fairlearn==0.10.0
faiss-cpu==1.7.4
fast-pareto==0.2.2
fasteners==0.18
fastjsonschema==2.18.0
filelock==3.12.3
flash-attn==1.0.5
fonttools==4.42.1
fqdn==1.5.1
frozenlist==1.4.0
fsspec==2023.10.0
future==0.18.3
gitdb==4.0.10
GitPython==3.1.32
gpytorch==1.11
graphviz==0.20.1
greenlet==2.0.2
gym==0.26.2
gym-notices==0.0.8
hyperopt==0.2.7
identify==2.5.27
idna==3.4
importlib-metadata==6.8.0
iniconfig==1.1.1
ipykernel==6.25.1
ipython==8.14.0
ipywidgets==8.1.0
iso8601==1.1.0
isoduration==20.11.0
jaxtyping==0.2.23
jedi==0.19.0
Jinja2==3.1.2
joblib==1.2.0
json5==0.9.14
jsonpath-ng==1.5.3
jsonpointer==2.4
jsonschema==4.19.0
jsonschema-specifications==2023.7.1
jupyter-events==0.7.0
jupyter-lsp==2.2.0
jupyter_ai==2.1.0
jupyter_ai_magics==2.1.0
jupyter_client==8.3.0
jupyter_core==5.3.1
jupyter_server==2.7.2
jupyter_server_terminals==0.4.4
jupyterlab==4.0.5
jupyterlab-pygments==0.2.2
jupyterlab-widgets==3.0.8
jupyterlab_server==2.24.0
kiwisolver==1.4.4
langchain==0.0.223
langchainplus-sdk==0.0.20
liac-arff==2.5.0
lightgbm==3.3.5
linear-operator==0.5.2
lit==16.0.6
llvmlite==0.39.1
locket==1.0.0
loguru==0.7.2
MarkupSafe==2.1.3
marshmallow==3.20.1
matplotlib==3.7.2
matplotlib-inline==0.1.6
memory-profiler==0.60.0
minio==7.1.17
mistune==3.0.1
mpmath==1.3.0
msgpack==1.0.5
multidict==6.0.4
multipledispatch==1.0.0
mypy-extensions==1.0.0
nbclient==0.8.0
nbconvert==7.7.4
nbformat==5.9.2
nest-asyncio==1.5.7
netaddr==0.8.0
networkx==3.1
ninja==1.11.1
nodeenv==1.8.0
notebook_shim==0.2.3
numba==0.56.4
numexpr==2.8.5
numpy==1.23.5
nvidia-cublas-cu11==11.10.3.66
nvidia-cublas-cu12==12.1.3.1
nvidia-cuda-cupti-cu11==11.7.101
nvidia-cuda-cupti-cu12==12.1.105
nvidia-cuda-nvrtc-cu11==11.7.99
nvidia-cuda-nvrtc-cu12==12.1.105
nvidia-cuda-runtime-cu11==11.7.99
nvidia-cuda-runtime-cu12==12.1.105
nvidia-cudnn-cu11==8.5.0.96
nvidia-cudnn-cu12==8.9.2.26
nvidia-cufft-cu11==10.9.0.58
nvidia-cufft-cu12==11.0.2.54
nvidia-curand-cu11==10.2.10.91
nvidia-curand-cu12==10.3.2.106
nvidia-cusolver-cu11==11.4.0.1
nvidia-cusolver-cu12==11.4.5.107
nvidia-cusparse-cu11==11.7.4.91
nvidia-cusparse-cu12==12.1.0.106
nvidia-ml-py3==7.352.0
nvidia-nccl-cu11==2.14.3
nvidia-nccl-cu12==2.18.1
nvidia-nvjitlink-cu12==12.3.52
nvidia-nvtx-cu11==11.7.91
nvidia-nvtx-cu12==12.1.105
openai==0.28.0
openapi-schema-pydantic==1.2.4
opencensus-context==0.1.3
openml==0.14.1
opt-einsum==3.3.0
oslo.concurrency==5.1.1
oslo.config==9.1.1
oslo.i18n==6.0.0
oslo.utils==6.1.0
osqp==0.6.5
overrides==7.4.0
packaging==23.1
pandas==2.0.3
pandocfilters==1.5.0
parso==0.8.3
partd==1.4.1
pathspec==0.11.2
pathtools==0.1.2
patsy==0.5.6
pexpect==4.8.0
pickleshare==0.7.5
Pillow==10.0.0
platformdirs==3.10.0
plotly==5.18.0
pluggy==1.0.0
ply==3.11
pre-commit==3.3.3
prometheus-client==0.17.1
prompt-toolkit==3.0.39
protobuf==4.24.1
psutil==5.9.3
ptyprocess==0.7.0
pure-eval==0.2.2
py-spy==0.4.0
py4j==0.10.9.7
pyarrow==18.1.0
pyasn1==0.6.1
pybind11==2.11.1
pycparser==2.21
pycryptodome==3.21.0
pydantic==1.10.12
pydot==2.0.0
Pygments==2.16.1
pygmo==2.19.5
pynisher==0.6.4
pyparsing==3.0.9
pyrfr==0.8.3
pyro-api==0.1.2
pyro-ppl==1.8.4
Pyro4==4.80
PySocks==1.7.1
pytest==7.2.0
python-dateutil==2.8.2
python-json-logger==2.0.7
pytz==2023.3
PyYAML==6.0
pyzmq==25.1.1
qdldl==0.1.7.post0
referencing==0.30.2
regex==2023.8.8
requests==2.32.3
rfc3339-validator==0.1.4
rfc3986==2.0.0
rfc3986-validator==0.1.1
rpds-py==0.9.2
safetensors==0.4.5
sas7bdat==2.2.3
SciencePlots==2.1.1
scikit-learn==1.2.1
scipy==1.11.4
scs==3.2.4.post1
seaborn==0.13.2
Send2Trash==1.8.2
sentencepiece==0.1.99
sentry-sdk==2.18.0
serpent==1.41
setproctitle==1.3.2
shap==0.46.0
shellingham==1.5.4
six==1.16.0
skorch==0.15.0
slicer==0.0.8
smac==1.2
smart-open==7.1.0
smmap==5.0.0
sniffio==1.3.0
sortedcontainers==2.4.0
soupsieve==2.4.1
spacy-legacy==3.0.12
spacy-loggers==1.0.5
SQLAlchemy==2.0.20
stack-data==0.6.2
statsmodels==0.14.1
stevedore==5.1.0
sympy==1.13.1
tabpfn==0.1.9
tabulate==0.9.0
tblib==3.0.0
tempeh==0.1.12
tenacity==8.2.3
tensorboard-data-server==0.7.2
terminado==0.17.1
text-unidecode==1.3
threadpoolctl==3.1.0
tiktoken==0.4.0
tinycss2==1.2.1
tomli==2.0.1
toolz==0.12.0
torch==1.13.1
torch-scatter==2.1.1+pt113cu117
torch-sparse==0.6.17+pt113cu117
torchaudio==0.13.1+cu117
torchmetrics==0.10.2
torchvision==0.14.1+cu117
tornado==6.3.3
tqdm==4.67.1
traitlets==5.9.0
triton==2.0.0
typeguard==2.13.3
typing-inspect==0.9.0
typing_extensions==4.9.0
tzdata==2023.3
uc-micro-py==1.0.3
uri-template==1.3.0
urllib3==2.0.4
virtualenv==20.24.3
wandb==0.18.5
wasabi==1.1.3
wcwidth==0.2.6
webcolors==1.13
webencodings==0.5.1
websocket-client==1.6.2
Werkzeug==3.1.3
widgetsnbextension==4.0.8
wrapt==1.15.0
xgboost==1.6.2
xmltodict==0.13.0
xxhash==3.5.0
yarl==1.9.2
zict==3.0.0
zipp==3.16.2